WIP Updating to match pso-java

+++ GameRunner.java
@@ -15,7 +15,7 @@ public class GameRunner {
      *
      * @return number of blocks cleared by the agent
      */
-    private static int run(ExtendedState state, Float[] weights) {
+    private static int run(ExtendedState state, double[] weights) {
         PlayerSkeleton player = new PlayerSkeleton();
 
         // Create new display frame only if display setting is true
@@ -43,9 +43,9 @@ public class GameRunner {
     }
     private static void parseWeights(String arg) {
         String[] args = arg.split(" ");
-        Float weights[] = new Float[args.length];
+        double weights[] = new double[args.length];
         for(int i = 0; i < args.length; i++) {
-            weights[i] = Float.parseFloat(args[i]);
+            weights[i] = Double.parseDouble(args[i]);
         }
         int result = run(new ExtendedState(), weights);
         Log.info(String.format("%d", result));
